Break: Revert to header-only design
