+++ include/simsimd/bitwise.h
@@ -21,6 +21,36 @@
 extern "C" {
 #endif
 
+simsimd_f32_t simsimd_sve_b1_hamming(simsimd_b1_t const* a, simsimd_b1_t const* b, simsimd_size_t d) {
+    simsimd_size_t i = 0;
+    svuint8_t d_vec = svdupq_n_u8(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0);
+    svbool_t pg_vec = svwhilelt_b8(i, d);
+    do {
+        svuint8_t a_vec = svld1_u8(pg_vec, a + i);
+        svuint8_t b_vec = svld1_u8(pg_vec, b + i);
+        svuint8_t a_xor_b_vec = sveor_u8_m(pg_vec, a_vec, b_vec);
+        d_vec = svadd_u8_m(pg_vec, d_vec, svcnt_u8_x(pg_vec, a_xor_b_vec));
+        i += svcntb() * __CHAR_BIT__;
+        pg_vec = svwhilelt_b32(i, d);
+    } while (svptest_any(svptrue_b32(), pg_vec));
+    return 1 - svaddv_u8(svptrue_b32(), d_vec);
+}
+
+simsimd_f32_t simsimd_avx512_b1_hamming(simsimd_b1_t const* a, simsimd_b1_t const* b, simsimd_size_t d) {
+    simsimd_size_t words = d / 128;
+    uint64_t const* a64 = (uint64_t const*)(a);
+    uint64_t const* b64 = (uint64_t const*)(b);
+    /// Contains 2x 64-bit integers with running population count sums.
+    __m128i d_vec = _mm_set_epi64x(0, 0);
+    for (simsimd_size_t i = 0; i != words; i += 2)
+        d_vec = _mm_add_epi64( //
+            d_vec,             //
+            _mm_popcnt_epi64(  //
+                _mm_xor_si128( //
+                    _mm_load_si128((__m128i const*)(a64 + i)), _mm_load_si128((__m128i const*)(b64 + i)))));
+    return _mm_cvtm64_si64(_mm_movepi64_pi64(d_vec)) + _mm_extract_epi64(d_vec, 1);
+}
+
 /**
  *  @brief  Optimized version for Tanimoto distance on @b exactly 166 bits,
  *          forming 21 incomplete bytes for the MACCS fingerprints used in
