+++ include/simsimd/spatial_batch.h
