+++ setup.py
@@ -13,9 +13,6 @@ __lib_name__ = "simsimd"
 compile_args = []
 link_args = []
 macros_args = []
-source_args = ["python/lib.c", "src/autovec.c"]
-libraries_args = []
-include_args = ["include", numpy.get_include()]
 
 if sys.platform == "linux":
     compile_args.append("-std=c11")
@@ -41,23 +38,11 @@ if sys.platform == "win32":
     compile_args.append("/fp:fast")
 
 
-arch = platform.machine()
-
-if arch == "x86_64":
-    libraries_args.append("simsimd_x86_avx2")
-    libraries_args.append("simsimd_x86_avx512")
-elif arch == "aarch64" or arch == "arm64":
-    libraries_args.append("simsimd_arm_neon")
-    libraries_args.append("simsimd_arm_sve")
-
-
 ext_modules = [
     Extension(
         "simsimd",
-        sources=source_args,
-        libraries=libraries_args,
-        library_dirs=["src"],
-        include_dirs=include_args,
+        sources=["python/lib.c"],
+        include_dirs=["include", numpy.get_include()],
         extra_compile_args=compile_args,
         extra_link_args=link_args,
         define_macros=macros_args,
